issue #11 - Create a DAO mock and 'use' method to DAO insert.
